pyproject.toml
[project]
name = "uzum"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.46",
    "celery>=5.6.2",
    "django>=6.0.2",
    "channels>=4.0.0",
    "channels-redis>=4.1.0",
    "argon2-cffi>=23.1.0",
    "elasticsearch>=8.15.0",
    "qrcode>=7.4.2",
    "django-celery-beat>=2.1.0",
    "django-ckeditor-5>=0.2.19",
    "django-filter>=25.2",
    "django-location-field>=2.7.3",
    "django-minio-storage>=0.5.9",
    "django-mptt>=0.18.0",
    "django-storages>=1.14.6",
    "django-unfold>=0.80.2",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "image>=1.5.33",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
    "ujson>=5.11.0",
    "django-elasticsearch-dsl>=9.0",
    "djangorestframework-stubs>=3.16.8",
]

[dependency-groups]
dev = [
    "blinker>=1.9.0",
    "selenium>=4.41.0",
    "selenium-wire>=5.1.0",
    "webdriver-manager>=4.0.2",
]
